Add files via upload
Here is what I managed to create for Review 2. I was struggling to properly read from the file, and write into a 2D array, however, that was mostly due me just waiting to long to start the assignment. I plan on working on this more tomorrow, and hopefully finishing it then.